(Mtc) Modify the way handle the re-INVITE.
The UE saves call type history if call type is not unknown when creates
`MtcSession` for correctly calculating a available call type when receives
an INVITE w/o SDP.
If a call type is changed while handles an INVITE w/o SDP, that changed call type will be handled within another transaction based on user's decision. plesae refer to a document in the bug.

Bug: b/336223107
Test: atest ImsStackNativeEnablerMtcTests, atest ImsStackNativeTests

Change-Id: I8fc086774011d6c3704458ce691e170a5143580e