(AoS) Modify the logic for setting common config
This commit modifies the logic for setting common Config.
According to ImsStackTestBase#writeTestConfig, if the Config is null,
the previous config will be deleted.
the logic is updated to reuse the member variable mConfig instead of using null in tests.
Plus, Added code to tearDown to initialize the MmTel capabilities for
each test using the initialized capabilities.

Bug: 318604770
Test: atest -c ImsStackTest:com.android.imsstack.its.tests.BasicRegistrationTest
Change-Id: I932361653726cfae779c7c4435e6e64eb14c9402