(AP_IMS)(Media) memory leak - deleting JNI params
When sending a message from Java to Native, JniMediaSession creates (new) the appropriate params and sends them to MediaManager.
MediaManager. then sends this message to its corresponding MediaSession for each method.
params are deleted after each method, but if an error occurs during this process, params are not deleted.
the role for deleting params to be moved to the MediaManager, not each media session.

Bug: 335790102
Test: MO/MT call
      atest ImsStackNativeEnablerMediaTests
Change-Id: Iff7ea8cc86bdbc9fead106eae010ad61d8e4e2f8