(Mtc) Copy ConfUsers before asynchronous calling ContinueConference
After IdleState#StartConference, MTC checks and waits the result of
radio operations such as startImsTraffic. This operation is done
asynchronously but MTC keeps using the references after the operation is
done. During this interval, the reference of ConfUser list becomes
pointing wrong information so there is an exception.
This changes IdleState to deep-copy the list so that it always points
valid objects.

Test: atest ImsStackNativeEnablerMtcTests
Bug: 341015539
Change-Id: Ifdcfb7b6ee5debded2750323a5682c3e159e687c