Fix incorrect display of hidden menu (2)
When entering ImsStack's hidden menu, there was a problem where the
first menu item was not visible due to the title bar or was located in
the position of the system bar.
This change is to position the first menu item below the title/system
bar so that the menu can be used normally.

Bug: 347786809
Flag: EXEMPT bugfix
Test: atest ImsStackJavaTests
Test: manual (verified IMS hidden menu on device)
Change-Id: I572a4ce781919cb2160daefa4f98985dee84c521