Fix a crash while displaying long log in x86 environment
A crash occurred when displaying log information longer that 512 in an
x86 environment. The first call to vsnprintf worked fine, but the second
call passed an already used va_list, resulting in incorrect string
formatting.
This change is to pass a copy of the va_list received as an input
argument to vsnprintf, preventing unexpected behavior depending on the
operating environment.

Bug: 349269417
Flag: EXEMPT bugfix
Test: atest ImsStackNativeTests
Test: atest ImsStackNativeEnablerAosTests:AosServiceTest#PrintCapabilities
Test: manual (verified IMS reg/call on device)
Change-Id: I4139ab5a9798375e0c4c9e41bb6564091e1d9feb