(Mtc) Adjust calling order of `SendUpdated` and `SendHeldBy` when receives up/downgrade, hold/resume simultaneously
When receives a INVITE request for up/downgrade, hold/resume
simultaneously, the current behavior is notifying the Telephony after
sending a 200 OK and receiving a ACK using the `callSessionHoldReceived`
with a changed call type and media direcions first and then
the `callSessionUpdated` with same info above.
It make sense that call the `SendUpdated` with changed info first
and then call the `SendHeldBy` because the `SendHeldBy` is only for
receiving Hold from peer instead of call type changing.

Flag: EXEMPT refactor
Bug: b/342104700
Test: atest ImsStackNativeEnablerMtcTests, atest ImsStackNativeTests,
verified using TISS.

Change-Id: I0ae24c784cad589193f563c254d78cf972c403c8