(Mtc) Fix for Crash Issue During Conference Call Termination
When a race condition occurs between the termination of conference call
and the user information update event(C-NOTIFY), accessing an already
removed IMtcCallContext results in a crash. This is a fundamental
structural issue with the ConferenceEventNotifier. In this fix, to
resolve the crash, the ConferenceEventNotifier will hold an
IMtcCallManager as a member instead of an IMtcCallContext, and use it to
ensure valid access to IJniMtcCallThread.

Test: simulation using cl/649227180
Test: atest ImsStackNativeTests
Bug: 349901867
Flag: NONE pre-production
Change-Id: I9104d0cc01a422f32cb1ff5b415b19faf22fc0b3