(Aos) Use GetInvokedCount instead of EXPECT_CALL
Test code written before the introduction of GetInvokedCount verified procedure execution by checking if a representative method within the procedure was called or not. This CL modifies to use GetInvokedCount instead of checking if a representative method.
And also removed some unnecessary checks of EXPECT_CALL.to focus on verifying specific results.

bug: 328362852
Test: atest ImsStackNativeTests
Flag: TEST_ONLY
Change-Id: Ic20bfbe624c4444793b25f0248441e6a04f95eb8