(Aos) fix the problem not to trigger registration
There is an issue where the registration attempt is not retried
if the registration is destroyed due to an IP change shortly
before receiving a response from startImsTraffic for registration transmission.
When registration is destroyed, update transaction block based on the
waiting flag of the current startImsTraffic callback.

Bug: 359436032
Test: manual and atest
Flag: EXEMPT pre-production
Change-Id: Ibe526e34743addd09ef191cbdb20cf3280cc8b39