(Aos) Fix the integer overflow on AosRegistration
There is an operation between signed and unsigned integers in AosRegistration and signed integer was implicitly converted to unsigned int.
The result of this operation was assigned to signed integer again and it caused integer overflow.
To prevent integer overflow, modified to explicitly transform the unsigned integer to signed integer before the operation.

Bug: 372097588
Test: atest ImsStackNativeTests with sanitize build option
Flag: NONE pre-production
Change-Id: I656e3be6511f97191c21c0f8149e7b55ed4270b6