(Mtc) SessionInterfaceHolder Improvement - CallKey Addition
Unlike other interfaces, the Session Interface may require information
tied to a specific call. This update enhances the Session Interface to
manage the CallKey when creating or adding a Session Interface. It also
improves the existing IInterfaceHolderListener so that the listener is
triggered when all ISession instances associated with the same CallKey
are cleared. Additionally, to allow other classes in MTC to register
listeners and receive information, the SessionInterfaceHolder will no
longer be created on-demand but will now follow the lifecycle of the
service.

Test: MO, MT, success and failure cases
Test: atest ImsStackNativeTests
Bug: 369458304
Flag: NONE pre-production
Change-Id: I0823ba917bf9e12cf7397ef86635f2895de1e3ce