Fix bug : call is not terminated in initiated state
The bug is side effect of ag/29798751.
If the call is terminated without receiving any provisional response,
including 100 Trying, the state remains in INITIATED. Currently,
ImsCallSessionImpl does not account for call termination in the
INITIATED state. Additionally, as various parts in ImsCallSessionImpl
are designed to function based on the NEGOTIATING state, this change
modifies it to avoid transitioning to the INITIATED state, just as it
was before.

Test: atest ImsStackJavaTests
Test: MO call rejected without 100 Trying
Test: MO call rejected with 100 Trying
Bug: 369458304
Flag: NONE pre-production
Change-Id: Iec7a495af4f3da8700c13d9d3cc2748918b33cdb