(Mtc) Send BYE if no ACK is received after accepting a call
When an incoming call is accepted and a 200 OK is sent, if no ACK is
received, the SIP engine invokes ISession#SessionStartFailed after the
200 OK retransmission timeout, transitioning the state to
STATE_ESTABLISHED. Currently, the Enabler terminates the call without
any additional SIP signaling, but the other party may already have the
call established. In this case, a BYE should be sent to properly end the
call. This modification ensures that a BYE is sent in such situations.

Test: atest ImsStackNativeMtcTests
Test: simulated the issue scenario
Bug: 376437749
Flag: NONE pre-production
Change-Id: Ia821c32ad66451a3bc75bd63adb3748d21b4202b