(Mtc) Check session state before sending PRACK
This change addresses an issue where a call disconnectes if the session
state transitions to STATE_ESTABLISHED before sending a PRACK for an
RPR. In MTC, SessionRprReceived() triggers a PRACK, but because the
Engine's session operates asynchronously, it may already have received a
200 OK, establishing the call by this point. If this happens, the Engine
returns a failure for SendPrack(), causing MTC to terminate the call. To
prevent this, MTC will verify if the Engine has already received the
final response before calling SendPrack(), avoiding unnecessary PRACK
attempts.

Test: simulated the issue scenario
Test: atest ImsStackNativeTests
Bug: b/376572567
Flag: NONE pre-production
Change-Id: I671c291abd48eba53521bc399929757093837f4a