(core) Fix 1xx response handling on session being terminated
While the 100 Trying response to INVITE was received and the session was
in process of being terminated (STATE_TERMINATING), a crash occurred
because an incorrect service method type was used during processing when
an 18x response was received.

When receiving a 1xx response, a service method type is selected
depending on the current session state, and the implementation has been
changed to process the 1xx response using this selected service method
type.

Bug: 376637921
Flag: EXEMPT bugfix
Test: manual (verify MO IMS call on device)
Test: manual (verify DS_EMC_037 test case)
Change-Id: Ia6b62675f8408d82e7e2181a36c02b551a3332de