(Mtc) Addition of EmergencyStartErrorHandler
This change introduces a new class, EmergencyStartErrorHandler, designed
to handle start failures specifically for emergency calls.
When a call is identified as an emergency call, the StartErrorHandler
will create an instance of EmergencyStartErrorHandler and delegate the
handling of the start failure to it.
Unless a carrier specifically requires otherwise, the CallReasonInfo
will be overridden with CODE_LOCAL_CALL_CS_RETRY_REQUIRED to allow for
retrying the emergency call.

Test: manual test with TISS
Test: atest ImsStackJavaTests
Test: atest ImsStackNativeTests
Bug: 364271399
Flag: NONE pre-production
Change-Id: I5d6664b49be8e048fd3abf7b21d4bc238c819f1f