(Aos) Add asset for capability when requesting IMS network
Regarding the network's VoPS indication, DataNetwork for IMS would be managed by Android Data Framework according to the Carrier Configurations.
IMS service can control to keep a DataNetwork regardless of the network's VoPS indication by requesting with IMS capability only.
This CL added an asset for this feature and enabled for the VZW.(CDMA-Less-Type3 TC 2.2.1.3 is failed if deactivate IMS PDN in VoPS not supported area)

Bug: 285467175
Test: atest ImsStackJavaTests and manual test
Flag: NONE pre-production
Change-Id: Ic173fc50af6ae96db7f88dd460a615da1ccf0e65