(Mtc) Prevent infinite loop of PendingOperation
If a timer for `TIMER_DELAY_UPDATE_AFTER_CONNECTED` is started and also
another pending operation is occurred during call establishing, infinite
loop is occurred when a call is established because
`MtcCall::RunPendingOperationIfPossible` pops and pushes a same
operation infinitely without changing a call state. So, no chance for
receiving expiration of that timer.

Bug: b/377858009
Test: atest ImsStackNativeEnablerMtcTests, simulated with the TISS
Flag: EXEMPT bugfix

Change-Id: Iba4a3c3181f3ca2754bb089b4267b0688caf2493