(SIP) Start Timer D regardless of transport protocol
Basically, in the case of an INVITE client transaction, after receiving
a failure response, it's decided whether to start Timer D according to
the SIP transport protocol - if UDP is used, Timer D is started,
otherwise, Timer D is not started.
However, the ACK request is transmitted by the SIP transaction layer,
and if the ACK request can't be transmitted immediately due to a problem
with the transport layer, it will be sent after a certain period of time
(TCP socket connection setup time).

This change is to start Timer D regardless of the SIP transport protocol
to properly handle this situation.

Bug: 373762975
Flag: EXEMPT bugfix
Test: atest ImsStackNativeTests
Test: manual (verify the transmission of ACK request using newly created
TCP socket)
Test: manual (verify IMS reg/call on device)

Change-Id: Ib48ccfc413a65aebc38d16e40435b9794efd5d90