(sipcore) Add socket recreation logic for INVITE client transaction
If the socket is already closed while transmitting an ACK request, the
Engine treats it as a transmission failure and there is no retry action.

This change is that if the socket is already closed while sending an ACK
request for an INVITE client transaction, the Engine recreates a new
socket and try to send ACK request using this newly created socket.

Bug: 373762975
Flag: EXEMPT bugfix
Test: atest ImsStackNativeTests
Test: manual (verify the transmission of ACK request using newly created
TCP socket)
Test: manual (verify IMS reg/call on device)

Change-Id: I0d438bcd98c570c673043cb4782d28b7d1ae1a30