(Mtc) Set Warning header by media negotiation failure result
This change ensures that when an incoming (re-)INVITE is rejected and
the CallReasonInfo code is set to CODE_MEDIA_NOT_ACCEPTABLE, the Warning
header is set based on the result of IMediaSession#NegotiateSdp. The
header is set in compliance with RFC 3261 Section 20.43, but due to the
limited information currently available from the MediaSession, only
partial values will be configured.

Test: atest ImsStackNativeEnablerMtcTests
Bug: 360734176
Flag: NONE pre-production
Change-Id: If2f19f5a70ebe3e923dddff4db773514695fd659