(Mtc) Reject re-INVITE(UPDATE) if media negotiation failed
This change ensures that when an incoming re-INVITE(or UPDATE) is
received and the result of IMediaSession#NegotiateSdp is not NO_ERROR,
the requeset will be rejected with a 488 response. Addtionally, the
negotiation result will be passed to the MessageSender so that an
appropriate Warning header can be set based on the outcome.

Test: atest ImsStackNativeEnablerMtcTests
Test: simulation the issue scenario and subsequent offer/answer
Bug: 360734176
Flag: NONE pre-production
Change-Id: I4ede5646117475bf8af276475584bee249f4de3b