Add Self-destroying with EnablerThread option to BaseService
This change prevents exceptions that could occur when BaseService is
deleted while it is asynchronously handling a message through
HandleMessage. The implementation ensures proper cleanup and thread
management to handle such scenarios gracefully.
The default behavior is to delete itself without thread switching and
the derived class should override Destroy() api if it needs to be
deleted using the enabler thread.

Test: stress test
Test: atest ImsStackNativeJniTests
Bug: 381020450
Flag: NONE pre-production
Change-Id: Ib3103ed9d7e83d7b66ca194e06a3817803b4422b