(Mtc) Change IsEpsCombinedAttach() and Add IsRoaming() to MtcService
This changes MtcService#IsEpsCombinedAttach to check only the LTE
infomation but the WLAN connection type. The callers must check WLAN
connection type by their desired result.
This also adds a new API IsRoaming() to MtcService.
This information is used for checking the supportablility of CS fallback
so there will be another API to check the unified CS fallback
supportability based on the configuration.

Test: atest ImsStackNativeEnablerMtcTests
Bug: 360045926
Flag: NONE pre-production
Change-Id: I08c7855c4d5209a6ce83ec4c1527c6b44f7028a1