(Mtc) Set a correct value of Terminal based call waiting
There is a no path for receiving a valid value of Terminal Based call waiting from the Telephony and the AP IMS Java layer has a default
value as false internally.
Even with this condition, it is okay for receiving a call waiting
indication because also the Native has a default value as UNPROVISIONED
internally among UNPROVISIONED, ENABLED, DISABLED and not declaring Terminal based call waiting feature in an internal default config file prevents setting a default false value from the Java layer.
However, the PARIS feature value is used for now and it makes disabling
Terminal based call waiting feature.
Therefore, the default value for Terminal based call waiting should be
set to true, as the AP IMS supports this functionality. If that value is
changed from the Telephony, it will be well reflected in the Native.

Bug: b/381019275
Flag: EXEMPT bugfix
Test: atest ImsStackJavaTests, atest ImsStackNativeEnablerMtcTests, simulating
Change-Id: I9f4ae496751cf7cd5ad7a4822638eba3b15be69d