(SIP) Fix incorrect decoding/encoding logic for Warning header
The current implementation of Warning header has the following problems:
 - Decode: No validation check for "warn-text".
 - Encode: Encoded the "warn-text" to the nested quoted-string if it
   already contains DQUOTE.

This change is to add a validation check for "warn-text" while decoding,
and use the decoded or the value that is set by the caller directly
while encoding. And, it also checks whether the input value of
SetWarnText(...) contains the DQUOTE and if not present,
SetWarnText(...) method adds the DQUOTE based on the syntax rule
accordingly.

Bug: 381160753
Flag: EXEMPT bugfix
Test: atest ImsStackNativeTests
Test: manual (verify the decoding/encoding of Warning header with TISS)
Test: manual (verify IMS reg/call on device)
Change-Id: I28af207c8f5a386a10d00fbf9484d689b9417c70