(Mtc) The call has to be barred even if the barring timer value is 0
The current implementation is not barring the call if a barring timer
value is 0 with a valid barring factor which can bar a call.
Based on the standard, it needs to be barred even if a barring timer
value is 0.
Please refer the standard summary in a b/374012247.

Bug: b/381788808
Flag: EXEMPT bugfix
Test: atest ImsStackNativeTests
Change-Id: Ic6f97adec95e4d5f32a1a68ff640036231094b1f