(plaform) Fix incorrect removal logic for already posted messages
When a JNI component is destroyed, if we no longer want to process
messages that have already been posted to the EnablerThread, we can
remove specific messages to prevent further processing. At this time,
among the messages currently being processed, the messages that have
already been processed may be included in the removal list, causing a
crash or message disappearance.

This change is to intended to correct crashes or message disappearance
by ensuring that already processed messages among messages that have
already passed to the process stage are not included in the removal
list.

Bug: 381826008
Flag: EXEMPT bugfix
Test: atest ImsStackNativeTests
Test: manual (verified no crashes or message disappearance not observed
using TISS)
Test: manual (verified IMS reg/call on device)

Change-Id: I452045fe03079127540c80764517cdc333bd5701