Use only valid subscription id for E-PDN setup
Basically, in a multi-SIM environment, the desired subscription id
should be passed to the connection request during PDN setup for normal
behavior.
However, in the case of an emergency call, E-PDN setup may be performed
even if there is no SIM or all SIMs are in LOCKED.
So this change is to prevent the invalid subscription id from being
passed when the subscription id is invalid.
This ensures that the Android Framework does not block E-PDN setup when
the subscription is invalid.

Bug: 380136223
Flag: EXEMPT bugfix
Test: atest ImsStackJavaTests
Test: manual (verified E-PDN setup when no SIM)
Change-Id: I2f996ce74b9a59faef812622c66adead0ff1be2a