Fix incorrect CS call state change
When making a normal VoLTE call in ECBM, there is a problem with the
phone state being recognized as a CS call.
This change is to ensure that when the IMS call state is updated to a
value other than the IDLE state, CS call state will be adjusted to go
into the IDLE state.

Bug: 382668074
Flag: EXEMPT bugfix
Test: atest ImsStackJavaTests
Test: manual (verified normal VoLTE call is successfully connected while
in ECBM)

Change-Id: Ia92557ab68afb2878db27ba29e6fbd78ea26917f