(Mtc) implement include_caller_id_service_codes_in_sip_invite_bool
This modification implements the behavior for the
include_caller_id_service_codes_in_sip_invite_bool configuration.
According to the definition of this configuration, if the dialed number
includes caller ID service codes such as *67 or *82, these codes are
removed. Additionally, the implementation addresses the handling of the
Privacy header values, which are implicitly required by this
configuration.
Note: the internal configuration values for default, TMO, AT&T, VZW are
not updated through this change since they are the public AOSP configs.
It seems the default value of CarrierConfigManager is also false so this
change can be applied for all carriers.

Test: atest ImsStackNativeEnablerMtcTests
Test: verify with config change and dialed number
Bug: 382332088
Flag: NONE pre-production
Change-Id: I4692468828f46da115f84f1911d92228c5e3cb07