(Mtc) Temporary change the QoS lost guard time as 5s
This change is a temporary fix to set the QoS lost guard time as 5
seconds. It is to extends to wait the QoS status between LTE and NR
handover.
But this is not the fundamental direction of the modification, so the
guard time value for QoS loss will be reverted to 1s.
A new timer will be considered to manage the QoS status due to handover
between LTE and NR.
Additionally, the internal timers of MTC related to QoS will be
extracted as carrier configuration items.

Test: Build, atest ImsStackNativeEnablerMtcTests
Bug: 382623945
Flag: NONE pre-production
Change-Id: I5da64fb13087a43682f62eb4c5c708bab1867434