(Mtc) OperationAsyncRunner is not deleted by RemoveCallback
This fixes an issue that the OperationAsyncRunner is not deleted by
RemoveCallbck in OperationAsyncRunnerManager. This changes to capture a
raw pointer in lambda to prevent dangling pointer. The current
implementation captures the 'pRunner' by value, which caused a dangling
pointer issue when the lambda was executed. This resulted in
'RemoveRunner' not being able to find and remove the corresponding
runner.
This commit captures the raw pointer of 'pRunner' instead, ensuring that
the lambda can access the correct object and 'RemoveRunner' can
successfully remove it.

Test: Check OpeartionAsyncRunner is deleted by callback
Test: atest ImsStackNativeEnablerMtcTests (no ut added, though)
Bug: 382144292
Flag: NONE pre-production
Change-Id: I13b80070f4cb31aa5f21b0cacc5c279b778206f6