(SIP) Fix an issue of failure return for a valid message
This change is to fix not updating the transaction status to Engine.
In INVITE server transaction, when received ACK request via TCP
for a failure response, transaction status was not updated.
Due to this failure returned to Engine. Moreover transaction was not terminated.
Fixed to update transaction status as valid message to Engine.

Bug: 382450182
Flag: EXEMPT bugfix
Test: atest ImsStackNativeTests
Test: manual (verified the issued testcase with TISS)
Test: manual (verified IMS reg/call on device)
Change-Id: Ia117ee6227acc70d069b4e1d2a1c2bcb72fe9c3b