(Mtc) Handling country specific URNs in 380 AS responses
When the device initiates a Non-UE detectable call, it may receive a 380
AS response from the network, which includes a country-specific URN.
In this case, the device should utilize the country-specific URN
provided in the Contact header of the 380 AS response for initiating
emergency calls.
This fix ensures that the country-specific URN is updated via the
ImsCallProfile before IMS relays the CODE_SIP_ALTERNATE_EMERGENCY_CALL
to Telephony. The transmitted URN is then relayed back when Telephony
requests an emergency call.

Test: manual test with TISS
Test: atest ImsStackJavaTests
Test: atest ImsStackNativeTests
Bug: 369455561
Flag: NONE pre-production
Change-Id: I20b74f6f93f25f1d3ccc6c2294da96c9b9168179