(Aos) Check if the network is available when emergency call ends
When the cellular network is lost, an emergency call is terminated and handover to IWLAN is performed.
But if the next emergency call is tried immediately, it can be sent again to the cellular network. To prevent this behavior, we modified it to check if the network is available when an emergency call ends.

Bug: 388576997
Test: atest ImsStackNativeTests
Flag: NONE pre-production
Change-Id: I1e7da54c0c967592fdce2c0cf29f653a0f1715aa