(Mtc) Refactor multiple dialog handling using MultipleDialogHandler
This modification introduces a MultipleDialogHandler class to handle the
behavior associated with multiple early dialogs within the
OutgoingState. Additionally, the responsibility for
RemoveInactiveSessions, previously managed by MtcCall, is transffered to
the MultipleDialogHandler.
In the future, this class will be extended to incorporate IR.92
compliance and operator-specific behaviors.

Test: atest ImsStackNativeEnablerMtcTests
Test: forking call case
Bug: 386126759
Flag: NONE pre-production
Change-Id: I11c8dcf54b3fcd0e9956d2509d598f6abff52e14