(engine) Support "reg-id" parameter in REGISTER request
This change is to add a new configuration for whether to support the
multiple registration feature, and it provides the configuration option
to either fully implement the multiple registration feature or use only
the "reg-id" parameter.
For now, only configuration that is not supported or uses the "reg-id"
parameter is applied.

Bug: 371453287
Flag: NONE pre-production
Test: atest ImsStackNativeTests
Test: manual (verified "reg-id" parameter is set or not set depending on
whether it is supported)
Test: manual (verified IMS reg/call on device)

Change-Id: Ia8449e8f2db272d58cf6c31f5dce48fe18b57073