(Mtc) fix crash during handling pending operation
This fixes a crash in EstablishedState::OnEnter if there is a pending
operation and it's handled synchronously. A pending operation here
normally changes the call state to UpdatingState so EstalishedState is
deleted directly. If there is more implementation after
RunPendingOperationIfPossible(), it might be going to make an exception.

Test: verified glare condition with 491
Test: ImsStackNativeEnablerMtcTests
Bug: 388210264
Flag: NONE pre-production
Change-Id: Ie79f61e45d709ff565d4d98b8ae6b8f57571da87