Use fresh/cached location for PIDF-LO on emergency call
This change is to ensure that the location is updated to the latest one
when dialing an emergency call for the carriers requesting Geolocation
information (PIDF-LO).
If a fresh location is not obtained, PIDF-LO is constructed based on the
last known location provided by LocationManager or cached location if
it's not expired yet.
To obtain the most recent location, the following trigger points are
added under specific conditions:
 - when ImsStack started
 - when carrier configuration is changed
 - when an outgoing emergency call is being initiated
 - when Wi-Fi setting is enabled

Bug: 384009148
Bug: 388625320
Flag: NONE pre-production
Test: atest ImsStackJavaTests
Test: manual (verified the refreshed or cached location when dialing an
emergency call)

Change-Id: I0bee4e00da4f2eb2725025e3926425132aa00d48