(Mtc) Add IMtcCallContext#IsEstablished
This IsEstablished() indicates that the call is in a confirmed state. In
other words, it returns true when the call is connected through an
INVITE/200/ACK sequence. Therefore, it will still return true even if
the call is in the UpdatgingState or TerminatingState after the call is
initially connected.

Test: atest ImsStackNativeEnablerMtcTests
Bug: 390327122
Flag: NONE pre-production
Change-Id: Ia3b5d1431d9642faf8ee1162281b841f77ef066a