(Mtc) Change reject code to 488 when MmTel feature is not available
Currently the UE rejects the incoming call with 406 if MmTel feature is
not available, in ServiceBlockRule.

It's originally from the NetworkBlockRule(see ag/30882236). But it's
different from the legacy solution that uses 488 and makes an AT&T TC
fail.

This CL changes the reject code to 488 to align with the legacy UCBlock
and AT&T requirement.

Bug: 386840390
Test: ImsStackNativeEnablerMtcTests
Test: TISS simulation with test code
Flag: NONE pre-production
Change-Id: I5d5cfb129d65fa10884b5324e20d7c848df11a4a