(sipcore) Close TCP client socket when transaction timeout occurs
This change is to close the TCP client socket used for the transaction
when a SIP transaction timeout occurs.
In order to avoid a situation where SIP message is delayed or cannot be
transmitted when another SIP transaction occurs during the TCP
retransmission process, when a SIP transaction timeout occurs, the SIP
transport layer closes the TCP client socket used for this transaction
so that a new SIP request can use the newly created socket.

Bug: 385008677
Flag: NONE pre-production
Test: atest ImsStackNativeTests
Test: manual (verified TCP client socket closed when SIP transaction
timeout occurs using TISS)
Test: manual (verified carrier test case)

Change-Id: I4c521e8bf29f46b9b709760ee1c14a28d5033893