(Mtc) Ensure WAIT_VIDEO_TEXT_AVAILABLE timer starts even w/o dedicated bearer wait timer
This change ensures the WAIT_VIDEO_TEXT_AVAILABLE timer starts during
the call setup phase, even when the Dedicated Bearer Wait timer is not
used.

Previously, without the WAIT_VIDEO_TEXT_AVAILABLE timer, an early UPDATE
would be sent as soon as the dedicated bearer for audio media was
acquired, setting the local status for video/text to 'none'. This change
uses the timer to ensure that the local status for all media, including
video/text, is updated in the UPDATE, as dedicated bearers for these
media may still be acquired even without the Dedicated Bearer Wait
timer.

Furthermore, this addresses a potential issue where a call could drop if
the dedicated bearer for video/text was acquired immediately after an
UPDATE was sent, as the ongoing transaction would prevent sending
another UPDATE. While further improvements are needed, this timer covers
most cases.

Test: atest ImsStackNativeEnablerMtcTests
Test: verified using simulation and test code
Bug: 382000664
Flag: NONE pre-production
Change-Id: Ifbc4908cafc9b8b51034dad66db2a5c1df2b1dd6