(Mtc) Add condition to immediately cancel emergency call
This change adds a condition to immediately cancel an emergency call and
stop retrying the connection if a specific SIP reject code and reason
header are received from the network after the call is initiated.
This behavior can be controlled through configuration, and currently,
this configuration is added for the AT&T carrier.
For the AT&T carrier, if a 503 response includes a specific reason
header, the message in the header will be displayed to the user, and the
emergency call will be canceled.

Test: manual test with TISS
Test: atest ImsStackNativeTests
Bug: 389560791
Flag: NONE pre-production
Change-Id: I4ad4357f21d70faf66c333f591875d67d0eb4211