(Aos) updated testcase when OnCallSessionReleased is invoked
Precondition:
- KEY_KEEP_EPDN_UPON_PCSCF_UNAVAILABLE_BOOL is true
- In case of OnCallSessionReleased is invoked without session establishment

do not release the PDN. (return)

E-PDN could be released when receiving CallTracker_ECallSessionReleased() according to ag/31187341
But There are cases that E-PDN should not be released if a session is not established.
If the session is not established, the E-PDN could be released by CallTracker_StateChanged().

The code is included. So I updated the testcase for this.

Flag: TEST_ONLY
Bug: 390504501
Test: atest
Change-Id: Ie8ae29c01d6c6df05002f1b0ac94c5f026d01814