(Mtc) Block or allow geolocation by PLMN when no UICC inserted
Verizon requests to send PIDF-LO even if there's no UICC inserted.

To support it with minimum impact, this commit adds a new config that
defines the list of PLMNs that allow PIDF-LO in INVITE if no UICC.

GEOLOCATON_BLOCK_CONDITION_ANONYMOUS is removed since the new config can
cover this.

Test: atest ImsStackNativeEnablerMtcTests
Test: TISS simulation and test code (REG_MODE_NOUICC)
Bug: 394205303
Flag: NONE pre-production
Change-Id: I339ad8cf6481b8bb3d9eb7c0b368f3e6463ef21c

Change-Id: I3ac183372c7d475b896a93ac2c297ea60ccc6618