(Aos) Modifies to register WifiWatcher observer only when WLAN is supported
This CL has been modified to register WifiWatcher observers only when WLAN rat is supported for the emergency type. For emergency type, Wifi connection status is used to release the emergency PDN when the WiFi network is disconnected. Therefore, we only need to register observers for operators that support WLAN rat.

Bug: 388576997
Test: atest ImsStackNativeTests
Flag: NONE pre-production
Change-Id: I2f49c74cdd3323ba85a3243d37ac20e68cb24b1b