(Mtc) Support SRVCC State update for emergency service
This change modifies MtcService with ServiceType::NORMAL triggers
UpdateSrvccState of MtcService with ServiceType::EMERGENCY to notify the
SRVCC state change to the MtcCalls with Emergency Service.
Since only MtcService with ServiceType::NORMAL sets the NativeEnabler
itself, it can only be notified with the SRVCC state change.
This change ensures all MtcCalls irrespective of the ServiceType can be
notified with the SRVCC state change.

Test: simulated using TISS
Test: atest ImsStackNativeEnablerMtcTests
Bug: 395735871
Flag: NONE pre-production
Change-Id: I44e594a9723fd104f14af1ba8e894d6c3d804c39