(core) Add new method to check incoming session cancellation
This change adds a new method needed to handle exceptional situations
when a glare condition occurs in SIP messages. This is to enable the
Enabler to control the SIP session by setting a flag when the device
receives an incoming INVITE request, sends 200 OK response, and then
receives an incoming CANCEL request.
Additionally, a new listener method - SessionCanceledOnAccepted is added
so that the Enabler can be notified immediately when the above situation
occurs.

Bug: 364408348
Flag: NONE pre-production
Test: atest ImsStackNativeTests
Test: manual (verified the flag is correctly set and listener method is
called when the glare condition occurs)

Change-Id: Idcc4558f638dde5d25fb17dbbbdec117ea81581b