(Aos) Register TelephonyCallback in case of no SIM
This CL modifies to register TelephonyCallback in no SIM case.
ImsService should receive PreciseDataConnectionState for the emergency service even though a SIM card is not inserted. In this case, we can register a TelephonyCallback with the default subId then receive PreciseDataConnectionState for phoneId 0 during emergency call processing.

Bug: 397762151
Test: atest ImsStackJavaTests and manual test
Flag: NONE pre-production
Change-Id: I2e4cd4b1c42baa64f00d0ef0bf845bd0b10dd084