(Mtc) Handles RRC connection reject
If the MTC receives ‘ImsRadio_OnConnectionFailed` response with a
`REASON_RRC_REJECT` reason in the LTE network, ignores it and waits
transaction timeout. If transaction timeout occurs and is not applicable
for CSFB(no CS or not CS combined attach), requests to the AoS for
initial REGISTER. The AoS sends `REGISTER`(might this is blocked by
modem side) and re-transmit if there is no response. A `REGISTER` is
actually sent from an UE after the amount of received time value in a
`ConnectionFailed`.
Another reason makes call termination.

Bug: 400557377
Test: atest ImsStackNativeEnablerMtcTests
Flag: EXEMPT bugfix
Change-Id: Ia6fbb5349c2144c6e5c642194e5245492f44ddd6