(Aos) added an interface to notify wifi setting to native
The UE needs to update to mtc depending on whether to do deregistration or not when wifi off.
Wifi setting status should be sent to native in order to be delivered to Mtc when IMS de-registration is needed.
(In case that handover is impossible. (no service, airplane mode, etc)

Flag: NONE pre-production
Bug: 397782832
Bug: 354052968
Test: manual test with Tiss and atest
Change-Id: I7484db8c5e9d0976aad1b68bc1bee5952f8ccb98