(Mtc) Dynamic emergency routing numbers
When an emergency call is initiated using a dynamic routing emergency
number, the AP IMS should be able to distinguish and process it as
either normal or emergency routing, depending on the IMS REG state.
Changes have been made to add dynamic routing emergency numbers to the
configuration. The AP IMS now verifies the IMS REG state for emergency
calls to these numbers. Additionally, the openEmergencyService call has
been repositioned to allow routing information to be determined based on
the requested number.

Test: manual test with TISS
Test: atest ImsStackJavaTests
Bug: 399029238
Flag: NONE pre-production
Change-Id: I0948ea8ce10b922e29bdb08b5ef76ec2429e8716