Update IMS call state to IDLE only when Telephony call state has changed
This CL is not to update the IMS call state to IDLE when Telephony
invokes onCallStateChanged() listener API with the same call state IDLE.
It may happen when ImsStack registers CallStateListener to Telephony
when the call starts. Then, Telephony notify the current call state and
it causes abnormal behaviors.

Test: atest PhoneStateAgentTest
Test: manual - dail call after device boot up
Bug: 385264520
Flag: NONE pre-production
Change-Id: Icfe91da592fec5a3a007c11fe1632695452afdcd