Config update for GCF - 100 trying timer to 0
This commit changes ims.sip_timer_100_trying_millis_int to 0.
The GCF test script expects the receiving device to always send a 100
Trying response, regardless of when the next response is transmitted.
This violates RFC 3261 and also contradicts 3GPP TS 34.229-1, which
states that the 100 Trying response is optional.
However, since convincing Anritsu would be a waste of time, we will
modify the relevant configuration on the device to support this
behavior.

Test: tested by chaning the configuration
Bug: 401440701
Flag: NONE pre-production
Change-Id: Ibb61d2ae2148ac0b95d9f26b5f2577f4854324a8