(Aos) Not to send initial subscription when the subscription is terminated.
Precondition
- ims.init_sub_upon_sub_terminated_bool: false (default value)
3GPP 24.229 5.1.1.3 Subscription to the registration-state event package
If a SUBSCRIBE request to refresh a subscription fails with a non-481 response, the UE shall still consider the original subscription valid for the duration of the most recently known "Expires" value according to RFC 6665 [28].
Otherwise, the UE shall consider the subscription invalid and start a new initial subscription according to RFC 6665 [28].
RFC 6665
A subscription is destroyed after a notifier sends a NOTIFY request
   with a "Subscription-State" of "terminated", or in certain error
   situations described elsewhere in this document.

Additionally, the UE used not to send initial subscription when the subscription is terminated when I checked the previous behavior.

Precondition
- ims.init_sub_upon_sub_terminated_bool: true
Upon expiration of the current subscription to the reg events package, the device shall initiate a new subscription request for the reg events package.

Flag: NONE pre-productiddon
Bug: 403434198
Test: manual test with Tiss and atest
Change-Id: Ida4fcce927f60df3feba9b478b95492474c3c92e