(platform) Update ICarrierConfig to check key presence for array type
This change is to update the related methods in ICarrierConfig to check
for the presence of a key for an array type.
The purpose of this is to allow API callers to properly handle various
config values by checking for the presence of a key via a second
argument.

Bug: 406876122
Flag: NONE pre-production
Test: atest ImsStackNativeTests
Test: manual (verified that 2nd argument value is set appropriately
based on the presence of the key)

Change-Id: I89e8e87663f6bbbed04a8044b4aef414d3aa1fea